// gauvendi-platform-service-v2/src/core/enums/common.ts
export enum HotelPaymentModeCodeEnum {
  NOGUAR = 'NOGUAR', // Reserve without credit card
  GUAWCC = 'GUAWCC', // Guarantee with credit card
  GUAWDE = 'GUAWDE', // Bank transfer
  GUAINV = 'GUAINV', // Pay with invoice
  PAYPAL = 'PAYPAL', //PayPal
  PMDOTH = 'PMDOTH', //Other Payment Method

  // TOTO remove
  CASH = 'CASH',
  BANK_TRANSFER = 'BANK_TRANSFER'
}
// ...
export enum ResponseCodeEnum {
  SUCCESS = '200',
  PENDING = '202',
  ERROR = '400'
}